New low-level API not based on iou.
This makes setting the user_data field private to ringbahn, so it
is safe to complete a CQE assuming it contains a Completion. With
some other changes around ring ownership, it makes the demo driver 100%
safe code.

The full iou API has not yet been implemented, just enough to implement
the demo driver.

The transition has begun to try to allow events to request multiple SQEs
from the SQ, allowing them to create linked events in a single go. The
first multi-SQE API will be the `hard_linked` API, which allows multiple
events to be linked together with `IORING_SQE_HARDLINK`. The results of
all but the final event will be ignored; the completion will only awaken
on the final event. This functionality is currently not correct
implemented.

+++ examples/read-event.rs
@@ -10,7 +10,7 @@ fn main() -> io::Result<()> {
     let submission = Submission::new(event, driver);
     let content = futures::executor::block_on(async move {
         let (event, result) = submission.await;
-        let bytes_read = result?;
+        let bytes_read = result? as usize;
         let s = String::from_utf8_lossy(&event.buf[0..bytes_read]).to_string();
         io::Result::Ok(s)
     })?;
